base: producer is not shared
self.producer is only used by the schema registry, not by the rest
proxy. So that instance doesn't need to be in the base.
